+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 
+## 0.20.4
 
 - updated dependencies: